refactor(ui): update layouts and components for i18n
- Fix Next.js layout nesting violations
- Update sidebar and top-nav for proper i18n
- Enhance button and separator components
- Add root layout updates
