+++ app/page.tsx
@@ -1,5 +1,6 @@
 import { redirect } from 'next/navigation';
+import { APP_ROUTES } from '../lib/routes';
 
 export default function RootPage() {
-  redirect('/es');
+  redirect(`/es/${APP_ROUTES.login}`);
 }
